add separate package for storage, sdd separate files for processing http requests
