+++ cmd/server/server_test.go
@@ -6,16 +6,16 @@ import (
 	"io"
 	"net/http"
 	"net/http/httptest"
-	"sync"
 	"testing"
 
 	"github.com/stretchr/testify/assert"
 	"github.com/stretchr/testify/require"
 	"go.uber.org/zap"
+
+	str "github.com/Tanya1515/metrics-collector.git/cmd/storage/structure"
 )
 
 func TestProcessRequest(t *testing.T) {
-	var mutex sync.Mutex
 	var gaugeMetricValue = 1.5
 	var counterMetrciValue int64 = 4
 	type httpResult struct {
@@ -27,7 +27,7 @@ func TestProcessRequest(t *testing.T) {
 		name    string
 		request string
 		metric  *Metrics
-		storage *MemStorage
+		storage *str.MemStorage
 		result  httpResult
 		modify  string
 	}{
@@ -35,7 +35,7 @@ func TestProcessRequest(t *testing.T) {
 			name:    "test: Send correct counter value",
 			request: "/update",
 			metric:  &Metrics{ID: "value", MType: "counter", Delta: &counterMetrciValue},
-			storage: &MemStorage{counterStorage: map[string]int64{"PollCount": 1}, gaugeStorage: map[string]float64{"BuckHashSys": 0.1}, mutex: &mutex},
+			storage: &str.MemStorage{},
 			result: httpResult{
 				code:        200,
 				response:    "{\"id\":\"value\",\"type\":\"counter\",\"delta\":4}",
@@ -47,7 +47,7 @@ func TestProcessRequest(t *testing.T) {
 			name:    "test: Send correct gauge value",
 			request: "/update",
 			metric:  &Metrics{ID: "value", MType: "gauge", Value: &gaugeMetricValue},
-			storage: &MemStorage{counterStorage: map[string]int64{"PollCount": 1}, gaugeStorage: map[string]float64{"BuckHashSys": 0.1}, mutex: &mutex},
+			storage: &str.MemStorage{},
 			result: httpResult{
 				code:        200,
 				response:    "{\"id\":\"value\",\"type\":\"gauge\",\"value\":1.5}",
@@ -59,7 +59,7 @@ func TestProcessRequest(t *testing.T) {
 			name:    "test: Send incorrect metric type",
 			request: "/update",
 			metric:  &Metrics{ID: "value", MType: "test", Delta: &counterMetrciValue},
-			storage: &MemStorage{counterStorage: map[string]int64{"PollCount": 1}, gaugeStorage: map[string]float64{"BuckHashSys": 0.1}, mutex: &mutex},
+			storage: &str.MemStorage{},
 			result: httpResult{
 				code:        400,
 				response:    "Error 400: Invalid metric type: test\n",
@@ -71,7 +71,7 @@ func TestProcessRequest(t *testing.T) {
 			name:    "test: Send none metric name",
 			request: "/update",
 			metric:  &Metrics{ID: "", MType: "counter", Delta: &counterMetrciValue},
-			storage: &MemStorage{counterStorage: map[string]int64{"PollCount": 1}, gaugeStorage: map[string]float64{"BuckHashSys": 0.1}, mutex: &mutex},
+			storage: &str.MemStorage{},
 			result: httpResult{
 				code:        404,
 				response:    "Error 404: Metric name was not found\n",
@@ -83,9 +83,15 @@ func TestProcessRequest(t *testing.T) {
 
 	for _, test := range tests {
 		t.Run("Test:", func(t *testing.T) {
+			err := test.storage.Init()
+			if err != nil {
+				panic(err)
+			}
+			test.storage.RepositoryAddCounterValue("PollCount", 1)
+			test.storage.RepositoryAddGaugeValue("BuckHashSys", 0.1)
 			var buf bytes.Buffer
 			bodyRequestEncode := json.NewEncoder(&buf)
-			err := bodyRequestEncode.Encode(test.metric)
+			err = bodyRequestEncode.Encode(test.metric)
 			if err != nil {
 				panic(err)
 			}
@@ -97,7 +103,7 @@ func TestProcessRequest(t *testing.T) {
 			}
 
 			defer logger.Sync()
-			App := Application{Storage: test.storage, logger: *logger.Sugar()}
+			App := Application{Storage: test.storage, Logger: *logger.Sugar()}
 
 			w := httptest.NewRecorder()
 
@@ -128,8 +134,6 @@ func TestProcessRequest(t *testing.T) {
 }
 
 func TestGetMetric(t *testing.T) {
-	var mutex sync.Mutex
-
 	type httpResult struct {
 		code        int
 		response    string
@@ -139,14 +143,14 @@ func TestGetMetric(t *testing.T) {
 		name    string
 		request string
 		metric  *Metrics
-		storage *MemStorage
+		storage *str.MemStorage
 		result  httpResult
 	}{
 		{
 			name:    "test: Send incorrect metric type",
 			request: "/value",
 			metric:  &Metrics{ID: "PollCount", MType: "test"},
-			storage: &MemStorage{counterStorage: map[string]int64{"PollCount": 1}, gaugeStorage: map[string]float64{"BuckHashSys": 0.1}, mutex: &mutex},
+			storage: &str.MemStorage{},
 			result: httpResult{
 				code:        400,
 				response:    "Error 400: Invalid metric type: test\n",
@@ -157,7 +161,7 @@ func TestGetMetric(t *testing.T) {
 			name:    "test: Get correct counter metric",
 			request: "/value",
 			metric:  &Metrics{ID: "PollCount", MType: "counter"},
-			storage: &MemStorage{counterStorage: map[string]int64{"PollCount": 1}, gaugeStorage: map[string]float64{"BuckHashSys": 0.1}, mutex: &mutex},
+			storage: &str.MemStorage{},
 			result: httpResult{
 				code:        200,
 				response:    "{\"id\":\"PollCount\",\"type\":\"counter\",\"delta\":1}",
@@ -168,7 +172,7 @@ func TestGetMetric(t *testing.T) {
 			name:    "test: Get correct gauge metric",
 			request: "/value",
 			metric:  &Metrics{ID: "BuckHashSys", MType: "gauge"},
-			storage: &MemStorage{counterStorage: map[string]int64{"PollCount": 1}, gaugeStorage: map[string]float64{"BuckHashSys": 0.1}, mutex: &mutex},
+			storage: &str.MemStorage{},
 			result: httpResult{
 				code:        200,
 				response:    "{\"id\":\"BuckHashSys\",\"type\":\"gauge\",\"value\":0.1}",
@@ -179,7 +183,7 @@ func TestGetMetric(t *testing.T) {
 			name:    "test: Get not existing gauge metric",
 			request: "/value",
 			metric:  &Metrics{ID: "GaugeTest", MType: "gauge"},
-			storage: &MemStorage{counterStorage: map[string]int64{"PollCount": 1}, gaugeStorage: map[string]float64{"BuckHashSys": 0.1}, mutex: &mutex},
+			storage: &str.MemStorage{},
 			result: httpResult{
 				code:        404,
 				response:    "Error 404: GaugeTest does not exist in gauge storage: ErrMetricExists\n",
@@ -190,7 +194,7 @@ func TestGetMetric(t *testing.T) {
 			name:    "test: Get not existing counter metric",
 			request: "/value",
 			metric:  &Metrics{ID: "PollCountEx", MType: "counter"},
-			storage: &MemStorage{counterStorage: map[string]int64{"PollCount": 1}, gaugeStorage: map[string]float64{"BuckHashSys": 0.1}, mutex: &mutex},
+			storage: &str.MemStorage{},
 			result: httpResult{
 				code:        404,
 				response:    "Error 404: PollCountEx does not exist in counter storage: ErrMetricExists\n",
@@ -201,7 +205,7 @@ func TestGetMetric(t *testing.T) {
 			name:    "test: Get request without metric name",
 			request: "/value",
 			metric:  &Metrics{ID: "", MType: "counter"},
-			storage: &MemStorage{counterStorage: map[string]int64{"PollCount": 1}, gaugeStorage: map[string]float64{"BuckHashSys": 0.1}, mutex: &mutex},
+			storage: &str.MemStorage{},
 			result: httpResult{
 				code:        404,
 				response:    "Error 404: Metric name was not found\n",
@@ -212,9 +216,15 @@ func TestGetMetric(t *testing.T) {
 
 	for _, test := range tests {
 		t.Run("Test:", func(t *testing.T) {
+			err := test.storage.Init()
+			if err != nil {
+				panic(err)
+			}
+			test.storage.RepositoryAddCounterValue("PollCount", 1)
+			test.storage.RepositoryAddGaugeValue("BuckHashSys", 0.1)
 			var buf bytes.Buffer
 			bodyRequestEncode := json.NewEncoder(&buf)
-			err := bodyRequestEncode.Encode(test.metric)
+			err = bodyRequestEncode.Encode(test.metric)
 			if err != nil {
 				panic(err)
 			}
@@ -226,7 +236,7 @@ func TestGetMetric(t *testing.T) {
 			}
 
 			defer logger.Sync()
-			App := Application{Storage: test.storage, logger: *logger.Sugar()}
+			App := Application{Storage: test.storage, Logger: *logger.Sugar()}
 
 			w := httptest.NewRecorder()
