+++ cmd/storage/structure/get_str_metric.go
@@ -1,46 +1,8 @@
-package main
+package storage
 
 import (
-	"sync"
-
 	"github.com/pkg/errors"
 )
-
-var errorMetricExists = errors.New("ErrMetricExists")
-
-type MemStorage struct {
-	counterStorage map[string]int64
-	gaugeStorage   map[string]float64
-	mutex          *sync.Mutex
-	backup         bool
-	fileStore      string
-}
-
-type ReposutiryInterface interface {
-	RepositoryAddCounterValue()
-	RepositoryAddGaugeValue()
-}
-
-func (S *MemStorage) RepositoryAddValue(metricName string, metricValue int64) {
-	S.mutex.Lock()
-
-	defer S.mutex.Unlock()
-	S.counterStorage[metricName] = metricValue
-}
-
-func (S *MemStorage) RepositoryAddCounterValue(metricName string, metricValue int64) {
-	S.mutex.Lock()
-
-	defer S.mutex.Unlock()
-	S.counterStorage[metricName] = S.counterStorage[metricName] + metricValue
-}
-
-func (S *MemStorage) RepositoryAddGaugeValue(metricName string, metricValue float64) {
-	S.mutex.Lock()
-
-	defer S.mutex.Unlock()
-	S.gaugeStorage[metricName] = metricValue
-}
 
 func (S *MemStorage) GetCounterValueByName(metricName string) (int64, error) {
 	S.mutex.Lock()
